feat: moved logic to postgres
