+++ src/jobs/index.ts
@@ -16,6 +16,7 @@ import "@/jobs/daily-volumes";
 import "@/jobs/update-attribute";
 import "@/jobs/collections-refresh";
 import "@/jobs/nft-balance-updates";
+import "@/jobs/data-export";
 
 // Export all job queues for monitoring through the UI.
 
@@ -57,7 +58,7 @@ import * as backfillNftBalanceFloorAskPriceQueue from "@/jobs/nft-balance-update
 import * as processActivityEvent from "@/jobs/activities/process-activity-event";
 import * as removeUnsyncedEventsActivities from "@/jobs/activities/remove-unsynced-events-activities";
 import * as fixActivitiesMissingCollection from "@/jobs/activities/fix-activities-missing-collection";
-import * as exportDataToS3 from "@/jobs/s3-export/export-data-to-s3";
+import * as exportData from "@/jobs/data-export/export-data";
 
 export const allJobQueues = [
   arweaveSyncBackfill.queue,
@@ -98,5 +99,5 @@ export const allJobQueues = [
   processActivityEvent.queue,
   removeUnsyncedEventsActivities.queue,
   fixActivitiesMissingCollection.queue,
-  exportDataToS3.queue,
+  exportData.queue,
 ];
