--- src/jobs/s3-export/index.ts
@@ -1,1 +1,0 @@
-import "@/jobs/s3-export/export-data-to-s3";
